requirements.txt
# RPA/AI Automation for Career College Enrollment Processing
# Brukd Consultancy - Process Automation Demo

# OCR and Document Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Email and Notifications
email-validator==2.1.0

# Web Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2

# AI/ML for Classification
scikit-learn==1.3.2
transformers==4.35.2

# Utilities
python-dotenv==1.0.0
reportlab==4.0.7
faker==20.1.0

